Level::pirateStationAction pass: restore signed index guard
linked 2756->2757, byte 1183->1183, avg 78.73->78.73; imports 596->596, extra 34->34.

+++ src/game/world/Level.cpp
@@ -469,7 +469,7 @@ void Level::pirateStationAction(bool param) {
         if (field_68 != 0) {
             return;
         }
-        if (Globals::status->getStation()->getPirateStationIndex() < 0) {
+        if ((int) Globals::status->getStation()->getPirateStationIndex() < 0) {
             return;
         }
         Array<bool> *tbl = Globals::status->field_4c;
